%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{aroma.tcga-package}
\alias{aroma.tcga-package}
\alias{aroma.tcga}
\docType{package}

\title{Package aroma.tcga}


 \description{
   Utility functions for accessing the public and protected data files at the The Cancer Genome Atlas (TCGA) Data Collection Center (DCC)

   This package should be considered to be in an alpha or beta phase.
   You should expect the API to be changing over time.
   Consider this package a private package.
 }

 \section{Installation and updates}{

   To install this package, see instructions at
   \url{https://r-forge.r-project.org/R/?group_id=429}.
 }

 \author{
  Henrik Bengtsson <henrikb@braju.com>.
 }

 \section{License}{
  LGPL (>= 2.1).
 }

 \references{
 
Please cite aroma.cn as

  H. Bengtsson, P. Neuvial and T.P. SpeedTumorBoost: Normalization of
  allele-specific tumor copy numbers from a single pair of tumor-normal
  genotyping microarrays, BMC Bioinformatics, 2010

  H. Bengtsson, A. Ray, P. Spellman and T.P. Speed. A single-sample
  method for normalizing and combining full-resolutioncopy numbers from
  multiple platforms, labs and analysis methods, Bioinformatics, 2009

  H. Bengtsson; K. Simpson; J. Bullard; K. Hansen. aroma.affymetrix: A
  generic framework in R for analyzing small to very large Affymetrix
  data sets in bounded memory, Tech Report 745, Department of
  Statistics, University of California, Berkeley, February 2008

  H. Bengtsson, R. Irizarry, B. Carvalho, & T.P. Speed. Estimation and
  assessment of raw copy numbers at the single locus level,
  Bioinformatics, 2008

 }
\keyword{package}
