%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  MultisourceCopyNumberChromosomalModel.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{fit.MultisourceCopyNumberChromosomalModel}
\alias{fit.MultisourceCopyNumberChromosomalModel}
\alias{MultisourceCopyNumberChromosomalModel.fit}
\alias{fit.MultisourceCopyNumberChromosomalModel}
\alias{fit,MultisourceCopyNumberChromosomalModel-method}


\title{Fits the model}

\description{
 Fits the model.
}

\usage{\method{fit}{MultisourceCopyNumberChromosomalModel}(...)}

\arguments{
  \item{...}{See subclasses.}
}

\value{
 See subclasses.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\seealso{
  For more information see \code{\link{MultisourceCopyNumberChromosomalModel}}.
}
\keyword{internal}
\keyword{methods}
