%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{dummy}
\alias{TcgaDccDownloader}
\alias{as.character.TcgaDccDownloader}
\alias{downloadCoreTextFile}
\alias{downloadCoreTextFile.TcgaDccDownloader}
\alias{downloadFile}
\alias{downloadFile.TcgaDccDownloader}
\alias{downloadUrl}
\alias{downloadUrl.TcgaDccDownloader}
\alias{getPassword}
\alias{getPassword.TcgaDccDownloader}
\alias{getProtocol}
\alias{getProtocol.TcgaDccDownloader}
\alias{getRootPath}
\alias{getRootPath.TcgaDccDownloader}
\alias{getRootUrl}
\alias{getRootUrl.TcgaDccDownloader}
\alias{getUrl.TcgaDccDownloader}
\alias{getUrlPath}
\alias{getUrlPath.TcgaDccDownloader}
\alias{getUrlPathname}
\alias{getUrlPathname.TcgaDccDownloader}
\alias{getUsername.TcgaDccDownloader}
\alias{listFiles}
\alias{listFiles.TcgaDccDownloader}
\alias{readDescription}
\alias{readDescription.TcgaDccDownloader}
\alias{readIdf}
\alias{readIdf.TcgaDccDownloader}
\alias{readLines}
\alias{readLines.default}
\alias{readLines.TcgaDccDownloader}
\alias{readManifest}
\alias{readManifest.TcgaDccDownloader}
\alias{readReadme}
\alias{readReadme.TcgaDccDownloader}
\alias{readSampleInfo}
\alias{readSampleInfo.TcgaDccDownloader}
\alias{readSdrf}
\alias{readSdrf.TcgaDccDownloader}
\alias{readTextFile}
\alias{readTextFile.TcgaDccDownloader}
\alias{setPassword}
\alias{setPassword.TcgaDccDownloader}
\alias{setUsername}
\alias{setUsername.TcgaDccDownloader}
\alias{downloadFiles}
\alias{downloadFiles.TcgaDccDownloader}
\alias{setLogin}
\alias{setLogin.TcgaDccDownloader}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
