%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  tcgaDccFullnameTranslator.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%


\name{tcgaDccFullnameTranslator.character}
\alias{tcgaDccFullnameTranslator.character}
\alias{tcgaDccFullnameTranslator.character}


\title{Translates a TCGA DCC fullname}

\description{
  Translates a TCGA DCC fullname into a comma separated format.
}

\usage{\method{tcgaDccFullnameTranslator}{character}(name, ..., set=NULL)}

\arguments{
 \item{name}{A \code{\link[base]{character}} string.}
 \item{...}{Not used.}
 \item{set}{...}
}

\value{
  Returns a \code{\link[base]{character}} string.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}

\references{
  [1] The TCGA Network, \emph{TCGA Data Primer - Version 1.0}, July 2008.\cr
}

\seealso{
  \code{\link[aroma.tcga:parseTcgaDccArchiveName.character]{*parseTcgaDccArchiveName}()}.
}


\keyword{methods}
\keyword{file}
