%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{TcgaDccDownloader}
\alias{as.character.TcgaDccDownloader}
\alias{downloadCoreTextFile}
\alias{downloadCoreTextFile.TcgaDccDownloader}
\alias{downloadFile}
\alias{downloadFile.TcgaDccDownloader}
\alias{downloadUrl}
\alias{downloadUrl.TcgaDccDownloader}
\alias{getPassword}
\alias{getPassword.TcgaDccDownloader}
\alias{getProtocol}
\alias{getProtocol.TcgaDccDownloader}
\alias{getRootPath}
\alias{getRootPath.TcgaDccDownloader}
\alias{getRootUrl}
\alias{getRootUrl.TcgaDccDownloader}
\alias{getUrl.TcgaDccDownloader}
\alias{getUrlPath}
\alias{getUrlPath.TcgaDccDownloader}
\alias{getUrlPathname}
\alias{getUrlPathname.TcgaDccDownloader}
\alias{getUsername.TcgaDccDownloader}
\alias{listFiles}
\alias{listFiles.TcgaDccDownloader}
\alias{readDescription}
\alias{readDescription.TcgaDccDownloader}
\alias{readIdf}
\alias{readIdf.TcgaDccDownloader}
\alias{readLines}
\alias{readLines.default}
\alias{readLines.TcgaDccDownloader}
\alias{readManifest}
\alias{readManifest.TcgaDccDownloader}
\alias{readReadme}
\alias{readReadme.TcgaDccDownloader}
\alias{readSampleInfo}
\alias{readSampleInfo.TcgaDccDownloader}
\alias{readSdrf}
\alias{readSdrf.TcgaDccDownloader}
\alias{readTextFile}
\alias{readTextFile.TcgaDccDownloader}
\alias{setPassword}
\alias{setPassword.TcgaDccDownloader}
\alias{setUsername}
\alias{setUsername.TcgaDccDownloader}
\alias{downloadFiles}
\alias{downloadFiles.TcgaDccDownloader}
\alias{setLogin}
\alias{setLogin.TcgaDccDownloader}
\alias{getAccessRoot}
\alias{getAccessRoot.TcgaDccDownloader}
\alias{getArchive}
\alias{getArchive.TcgaDccDownloader}
\alias{getCenter}
\alias{getCenter.TcgaDccDownloader}
\alias{getCenterType}
\alias{getCenterType.TcgaDccDownloader}
\alias{getDataType}
\alias{getDataType.TcgaDccDownloader}
\alias{getPlatform.TcgaDccDownloader}
\alias{getTumorType}
\alias{getTumorType.TcgaDccDownloader}
\alias{getHeader.MageTabDataMatrixFile}
\alias{getReadArguments.MageTabDataMatrixFile}
\alias{MageTabDataMatrixFile}
\alias{MageTabFile}
\alias{getExtractNameByFilename}
\alias{getExtractNameByFilename.SdrfFile}
\alias{getFilenameMap}
\alias{getFilenameMap.SdrfFile}
\alias{parseSdrfExtractName}
\alias{pruneSDRF}
\alias{readSDRFFile}
\alias{readSDRFFolder}
\alias{SdrfFile}
\alias{parseTcgaDccArchiveName}
\alias{tcgaDccFullnameTranslator}
\alias{BroadAlleleSpecificCopyNumberTcgaDataFile}
\alias{BroadAlleleSpecificThetaTcgaDataFile}
\alias{extractAlleleSpecificCopyNumbers}
\alias{extractAlleleSpecificCopyNumbers.BroadAlleleSpecificCopyNumberTcgaDataFile}
\alias{extractTheta}
\alias{extractTheta.BroadAlleleSpecificThetaTcgaDataFile}
\alias{extractTotalAndFracB}
\alias{extractTotalAndFracB.BroadAlleleSpecificCopyNumberTcgaDataFile}
\alias{extractTotalAndFracB.BroadAlleleSpecificThetaTcgaDataFile}
\alias{getFilenamePattern}
\alias{getFilenamePattern.BroadAlleleSpecificCopyNumberTcgaDataFile}
\alias{getFilenamePattern.BroadAlleleSpecificThetaTcgaDataFile}
\alias{getFilenamePattern.TcgaDataFile}
\alias{getReadArguments.BroadAlleleSpecificCopyNumberTcgaDataFile}
\alias{getReadArguments.BroadAlleleSpecificThetaTcgaDataFile}
\alias{TcgaDataFile}
\alias{BroadTotalCopyNumberTcgaDataFile}
\alias{extractTotalAndFracB.BroadTotalCopyNumberTcgaDataFile}
\alias{extractTotalCopyNumbers}
\alias{extractTotalCopyNumbers.BroadTotalCopyNumberTcgaDataFile}
\alias{getFilenamePattern.BroadTotalCopyNumberTcgaDataFile}
\alias{getReadArguments.BroadTotalCopyNumberTcgaDataFile}
\alias{makeFullNamesTranslator}
\alias{makeFullNamesTranslator.SdrfFile}
\alias{setFullNamesTranslatorBySdrfFile}
\alias{setFullNamesTranslatorBySdrfFile.GenericDataFileSet}
\alias{byPath.SdrfFileSet}
\alias{makeFullNamesTranslator.SdrfFileSet}
\alias{SdrfFileSet}
\alias{setFullNamesTranslatorBySdrfFileSet}
\alias{setFullNamesTranslatorBySdrfFileSet.GenericDataFileSet}
\alias{BiospecimenCoreResource}
\alias{findDataSets}
\alias{findDataSets.TcgaDccData}
\alias{getBarcodePattern}
\alias{getBarcodePattern.BiospecimenCoreResource}
\alias{getBarcodePatterns}
\alias{getBarcodePatterns.BiospecimenCoreResource}
\alias{getDataSetPattern}
\alias{getDataSetPattern.TcgaDccData}
\alias{getDataSetPatterns}
\alias{getDataSetPatterns.TcgaDccData}
\alias{getFullNameTranslator.MscnExport}
\alias{MscnExport}
\alias{TcgaDccData}
\alias{findDataSets2}
\alias{findDataSets2.TcgaDccData}
\alias{hasFile.TcgaDccDownloader}
\alias{hasFiles}
\alias{hasFiles.TcgaDccDownloader}
\alias{getKnownFilePatterns}
\alias{getKnownFilePatterns.TcgaDccDownloader}
\alias{getDataSet}
\alias{getDataSet.TcgaDccDownloader}
\alias{getDataSetName}
\alias{getDataSetName.TcgaDccDownloader}
\alias{downloadFilesByPattern}
\alias{downloadFilesByPattern.TcgaDccDownloader}
\alias{appendFullNamesTranslatorBySdrfFile}
\alias{appendFullNamesTranslatorBySdrfFile.GenericDataFileSet}
\alias{appendFullNamesTranslatorBySdrfFileSet}
\alias{appendFullNamesTranslatorBySdrfFileSet.GenericDataFileSet}
\alias{getBarcodeIdTranslator}
\alias{getBarcodeIdTranslator.TcgaFullNameTranslators}
\alias{getBarcodeIdTranslators}
\alias{getBarcodeIdTranslators.TcgaFullNameTranslators}
\alias{TcgaFullNameTranslators}
\alias{appendFullNameTranslatorBySdrfFile}
\alias{appendFullNameTranslatorBySdrfFile.FullNameInterface}
\alias{appendFullNameTranslatorBySdrfFileSet}
\alias{appendFullNameTranslatorBySdrfFileSet.FullNameInterface}
\alias{getExtensionPattern.BroadAlleleSpecificCopyNumberTcgaDataFile}
\alias{getExtensionPattern.BroadAlleleSpecificThetaTcgaDataFile}
\alias{getExtensionPattern.BroadTotalCopyNumberTcgaDataFile}
\alias{getExtensionPattern.TcgaDataFile}
\alias{BroadBirdseedGenotypeTcgaDataFile}
\alias{BroadBirdseedGenotypeTcgaDataFileSet}
\alias{byPath.MageTabFileSet}
\alias{extractCalls.BroadBirdseedGenotypeTcgaDataFile}
\alias{extractConfidenceScores}
\alias{extractConfidenceScores.BroadBirdseedGenotypeTcgaDataFile}
\alias{getExtensionPattern.BroadBirdseedGenotypeTcgaDataFile}
\alias{getReadArguments.BroadBirdseedGenotypeTcgaDataFile}
\alias{MageTabFileSet}
\alias{BroadAlleleSpecificCopyNumberTcgaDataFileSet}
\alias{BroadAlleleSpecificThetaTcgaDataFileSet}
\alias{BroadTotalCopyNumberTcgaDataFileSet}
\alias{MageTabDataMatrixFileSet}
\alias{TcgaDataFileSet}
\alias{exportGenotypeCallsAndConfidenceScores}
\alias{getFilenameMap.SdrfFileSet}
\alias{exportTotalAndFracB}
\alias{exportTotalAndFracB.BroadTotalCopyNumberTcgaDataFile}
\alias{exportTotalAndFracB.BroadTotalCopyNumberTcgaDataFileSet}
\alias{exportTotalAndFracB.HudsonAlphaBAlleleFreqTcgaDataFile}
\alias{exportTotalAndFracB.HudsonAlphaBAlleleFreqTcgaDataFileSet}
\alias{extractFracB}
\alias{extractFracB.HudsonAlphaBAlleleFreqTcgaDataFile}
\alias{getExtensionPattern.HudsonAlphaBAlleleFreqTcgaDataFile}
\alias{getReadArguments.HudsonAlphaBAlleleFreqTcgaDataFile}
\alias{HudsonAlphaBAlleleFreqTcgaDataFile}
\alias{HudsonAlphaBAlleleFreqTcgaDataFileSet}
\alias{exportGenotypeCallsAndConfidenceScores.BroadBirdseedGenotypeTcgaDataFileSet}
\alias{exportTotalAndFracB.HudsonAlphaXYTcgaDataFile}
\alias{exportTotalAndFracB.HudsonAlphaXYTcgaDataFileSet}
\alias{extractTheta.HudsonAlphaXYTcgaDataFile}
\alias{extractTotalAndFracB.HudsonAlphaXYTcgaDataFile}
\alias{getExtensionPattern.HudsonAlphaXYTcgaDataFile}
\alias{getReadArguments.HudsonAlphaXYTcgaDataFile}
\alias{HudsonAlphaXYTcgaDataFile}
\alias{HudsonAlphaXYTcgaDataFileSet}
\alias{exportGenotypeCalls}
\alias{exportGenotypeCalls.HudsonAlphaGenotypeTcgaDataFileSet}
\alias{extractCalls.HudsonAlphaGenotypeTcgaDataFile}
\alias{getExtensionPattern.HudsonAlphaGenotypeTcgaDataFile}
\alias{getReadArguments.HudsonAlphaGenotypeTcgaDataFile}
\alias{HudsonAlphaGenotypeTcgaDataFile}
\alias{HudsonAlphaGenotypeTcgaDataFileSet}
\alias{exportTotal}
\alias{exportTotal.HarvardTotalCopyNumberTcgaDataFile}
\alias{exportTotal.HarvardTotalCopyNumberTcgaDataFileSet}
\alias{extractTotalCopyNumbers.HarvardTotalCopyNumberTcgaDataFile}
\alias{extractTotalLog2CopyNumbers}
\alias{extractTotalLog2CopyNumbers.HarvardTotalCopyNumberTcgaDataFile}
\alias{getReadArguments.HarvardTotalCopyNumberTcgaDataFile}
\alias{HarvardTotalCopyNumberTcgaDataFile}
\alias{HarvardTotalCopyNumberTcgaDataFileSet}
\alias{exportTotal.MskccTotalCopyNumberTcgaDataFile}
\alias{exportTotal.MskccTotalCopyNumberTcgaDataFileSet}
\alias{extractTotalCopyNumbers.MskccTotalCopyNumberTcgaDataFile}
\alias{extractTotalLog2CopyNumbers.MskccTotalCopyNumberTcgaDataFile}
\alias{getReadArguments.MskccTotalCopyNumberTcgaDataFile}
\alias{MskccTotalCopyNumberTcgaDataFile}
\alias{MskccTotalCopyNumberTcgaDataFileSet}
\alias{getDefaultFullName.TcgaDataFileSet}
\alias{BroadAfter5NNTotalCopyNumberTcgaDataFile}
\alias{BroadAfter5NNTotalCopyNumberTcgaDataFileSet}
\alias{exportTotalAndFracB.BroadAfter5NNTotalCopyNumberTcgaDataFileSet}
\alias{getExtensionPattern.BroadAfter5NNTotalCopyNumberTcgaDataFile}
\alias{getColumnNames.MskccTotalCopyNumberTcgaDataFile}
\alias{extractByTcgaType}
\alias{extractByTcgaType.GenericDataFileSet}
\alias{readBCRSamplesTable}
\alias{readBCRSamplesTable.TcgaDccDownloader}
\alias{extractTumorNormalPairs}
\alias{extractTumorNormalPairs.GenericDataFileSet}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
